test(kubectl_manifest): migrate acc suite to framework_test
Move the 27 kubectl_manifest acceptance tests off the SDK v2 testing
harness and onto the muxed provider's ProtoV6ProviderFactories so they
exercise the framework half once the cutover (#62) flips the dispatch
target. Closes the first of three acceptance criteria on #295.

Layout changes:

- internal/framework/resource_kubectl_manifest_acc_test.go: 26 TestAcc*
  functions plus TestKubectlManifest_RetryOnFailure and TestAcckubectlYaml
  (moved from kubernetes/resource_kubectl_examples_test.go). All use
  ProtoV6ProviderFactories: testAccProtoV6ProviderFactories defined
  alongside the existing data-source and ephemeral acc tests.
- internal/framework/resource_kubectl_manifest_acc_helpers_test.go:
  testAccCheckkubectlDestroy / Exists / Status (rewritten to build a
  Kubernetes clientset directly from KUBECONFIG, since the muxed
  provider's configured KubeProvider is not exposed through the
  terraform-plugin-testing harness) plus the nginx fixture helpers.
- kubernetes/get_live_manifest_fields_test.go: the
  TestGetLiveManifestFilteredForUserProvidedOnly unit test plus its
  two helpers (withAlteredField, loadRealDeploymentManifest), kept in
  the kubernetes package because they exercise package-private
  GetLiveManifestFields / GetFingerprint.

Files deleted: kubernetes/resource_kubectl_manifest_test.go,
kubernetes/resource_kubectl_manifest_fixtures_test.go,
kubernetes/resource_kubectl_examples_test.go.

The SDK v2 imperative-check helpers (testAccCheckkubectl* in
kubernetes/provider_test.go) stay in place; the data source acc tests
in kubernetes/data_source_kubectl_manifest_test.go still use them and
will migrate with the data source itself in a separate task.

Refs: alekc/terraform-provider-kubectl#295
