+++ kubernetes/get_live_manifest_fields_test.go
@@ -2,1020 +2,14 @@ package kubernetes
 
 import (
 	"bytes"
-	"fmt"
 	"log"
 	"os"
-	"regexp"
 	"testing"
 
 	"github.com/alekc/terraform-provider-kubectl/yaml"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/stretchr/testify/assert"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
-
-func TestKubectlManifest_RetryOnFailure(t *testing.T) {
-	_ = os.Setenv("KUBECTL_PROVIDER_APPLY_RETRY_COUNT", "5")
-
-	config := `
-resource "kubectl_manifest" "test" {
-	yaml_body = <<YAML
-apiVersion: v1
-kind: Ingress
-YAML
-}
-	`
-
-	expectedError, _ := regexp.Compile(".*failed to create kubernetes.*")
-	resource.Test(t, resource.TestCase{
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				ExpectError: expectedError,
-				Config:      config,
-			},
-		},
-	})
-}
-
-func TestAccKubectl(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxPodConfig(`  wait = true`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_Wait(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxDeploymentConfig(`  wait = true`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_WaitBackground(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxDeploymentConfig(`
-  wait = true
-  delete_cascade = "Background"
-`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForRolloutDeployment(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxDeploymentConfigReplicas(`  wait_for_rollout = true`, 3)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForRolloutDaemonSet(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxDaemonSetConfig(`  wait_for_rollout = true`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-// TestAccKubectl_WaitForRolloutDaemonSetNoMatchingNodes is a regression test
-// for https://github.com/alekc/terraform-provider-kubectl/issues/228.
-// A DaemonSet whose nodeSelector matches no nodes settles with
-// DesiredNumberScheduled = 0 — the controller has nothing to schedule. Prior
-// to the fix, the rollout-wait opened a Watch with no initial-state probe;
-// the controller had usually already settled by then, no further Modified
-// events were emitted, and the apply blocked until Terraform's create
-// timeout. The tight `timeouts { create = "60s" }` budget below would have
-// failed under the old behaviour.
-func TestAccKubectl_WaitForRolloutDaemonSetNoMatchingNodes(t *testing.T) {
-	//language=hcl
-	config := `
-resource "kubectl_manifest" "test" {
-  wait_for_rollout = true
-  timeouts {
-    create = "60s"
-  }
-  yaml_body = <<YAML
-apiVersion: apps/v1
-kind: DaemonSet
-metadata:
-  name: nginx-noselect
-  labels:
-    app: nginx-noselect
-spec:
-  updateStrategy:
-    type: RollingUpdate
-  selector:
-    matchLabels:
-      app: nginx-noselect
-  template:
-    metadata:
-      labels:
-        app: nginx-noselect
-    spec:
-      nodeSelector:
-        terraform-provider-kubectl-test/no-such-label: "true"
-      containers:
-        - name: nginx
-          image: registry.k8s.io/e2e-test-images/nginx:1.28.0-1
-          ports:
-            - containerPort: 80
-YAML
-}
-`
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-			},
-		},
-	})
-}
-
-// TestAccKubectl_WaitForRolloutDeploymentUpdate is a regression test for
-// https://github.com/alekc/terraform-provider-kubectl/issues/226. Two
-// behaviours converged in that issue:
-//   - the resource schema only declared a `create` timeout, so users
-//     setting `timeouts { update = ... }` were silently ignored, and
-//   - `waitForDeploymentRollout` opened a Watch without an initial-state
-//     probe; an update whose rollout settled between the spec write and
-//     the Watch open then blocked until the operation timeout.
-//
-// This test exercises both: it creates a Deployment, then runs an
-// in-place update with an `update` timeout configured. The new code
-// path returns immediately once the rollout settles; the old path
-// hung 20 minutes and failed.
-func TestAccKubectl_WaitForRolloutDeploymentUpdate(t *testing.T) {
-	t.Parallel()
-
-	// Generate one unique name per test run so this test does not
-	// collide with concurrent runs (matrix jobs sharing a cluster,
-	// or other `t.Parallel()` tests that touch the default namespace).
-	name := acctest.RandomWithPrefix("issue-226-deployment")
-
-	deploymentConfig := func(name string, replicas int) string {
-		return fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-  wait_for_rollout = true
-  timeouts {
-    create = "60s"
-    update = "60s"
-    delete = "60s"
-  }
-  yaml_body = <<YAML
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: %s
-  labels:
-    app: %s
-spec:
-  replicas: %d
-  selector:
-    matchLabels:
-      app: %s
-  template:
-    metadata:
-      labels:
-        app: %s
-    spec:
-      containers:
-        - name: pause
-          image: registry.k8s.io/pause:3.10
-YAML
-}
-`, name, name, replicas, name, name)
-	}
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: deploymentConfig(name, 1)},
-			{Config: deploymentConfig(name, 2)},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForRolloutStatefulSet(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxStatefulSetConfig(`  wait_for_rollout = true`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_RequireWaitForFieldOrCondition(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxPodConfig(`  wait_for { }`)
-	expectedError := regexp.MustCompile(".*at least one of `field` or `condition` must be provided in `wait_for` block.*")
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      config,
-				ExpectError: expectedError,
-			},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForNegativeField(t *testing.T) {
-	t.Parallel()
-
-	ns := acctest.RandomWithPrefix("wait-for-neg-field")
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test_wait_for" {
-  timeouts {
-    create = "10s"
-  }
-  yaml_body = <<EOF
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: %s
-EOF
-
-  wait_for {
-    field {
-      key = "status.phase"
-      value = "Activez"
-    }
-  }
-}
-`, ns)
-	errorRegex := regexp.MustCompile(".*Wait returned an error*")
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      config,
-				ExpectError: errorRegex,
-			},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForNegativeCondition(t *testing.T) {
-	t.Parallel()
-
-	name := acctest.RandomWithPrefix("busybox-sleep")
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	timeouts {
-		create = "20s"
-	}
-
-	wait_for {
-		condition {
-			type = "ContainersReady"
-			status = "Never"
-		}
-	}
-	yaml_body = <<YAML
-apiVersion: v1
-kind: Pod
-metadata:
-  name: %s
-spec:
-  containers:
-  - name: busybox
-    image: busybox
-    command: ["sleep", "30"]
-YAML
-}
-`, name)
-	errorRegex := regexp.MustCompile(".*Wait returned an error*")
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      config,
-				ExpectError: errorRegex,
-			},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForNS(t *testing.T) {
-	t.Parallel()
-
-	ns := acctest.RandomWithPrefix("wait-for-ns")
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test_wait_for" {
-  timeouts {
-    create = "200s"
-  }
-  yaml_body = <<EOF
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: %s
-EOF
-
-  wait_for {
-    field {
-      key = "status.phase"
-      value = "Active"
-    }
-  }
-}
-`, ns)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForField(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxPodConfig(`
-	wait_for {
-		field {
-			key = "status.containerStatuses.[0].ready"
-			value = "true"
-		}
-		field {
-			key = "status.phase"
-			value = "Running"
-		}
-		field {
-			key = "status.podIP"
-			value = "^(\\d+(\\.|$)){4}"
-			value_type = "regex"
-		}
-	}
-`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForConditions(t *testing.T) {
-	t.Parallel()
-
-	config, _ := nginxPodConfig(`
-	wait_for {
-		condition {
-			type = "ContainersReady"
-			status = "True"
-		}
-		condition {
-			type = "Ready"
-			status = "True"
-		}
-	}
-`)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: config},
-		},
-	})
-}
-
-// deploymentUpdateConfig is a step-pair fixture for the Wait*Update tests:
-// both create and update configs target the SAME Deployment name (otherwise
-// step 2 would create a new Deployment, not update the existing one) and
-// share the same template. The caller supplies the kubectl_manifest-level
-// extra args for each step (e.g. wait_for_rollout = true on create, wait_for
-// { ... } block on update) and the desired replicas in the update step.
-func deploymentUpdateConfig(createExtra, updateExtra string, updateReplicas int) (createCfg, updateCfg string) {
-	name := acctest.RandomWithPrefix("nginx-deployment")
-	tmpl := func(extra string, replicas int) string {
-		return fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-%s
-	yaml_body = <<YAML
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: %s
-  labels:
-    app: %s
-spec:
-  replicas: %d
-  selector:
-    matchLabels:
-      app: %s
-  template:
-    metadata:
-      labels:
-        app: %s
-    spec:
-      containers:
-        - name: nginx
-          image: %s
-          ports:
-            - containerPort: 80
-          readinessProbe:
-            httpGet:
-              path: "/"
-              port: 80
-            initialDelaySeconds: 10
-YAML
-}
-`, extra, name, name, replicas, name, name, nginxImage)
-	}
-	return tmpl(createExtra, 1), tmpl(updateExtra, updateReplicas)
-}
-
-func TestAccKubectl_WaitForConditionUpdate(t *testing.T) {
-	t.Parallel()
-
-	createConfig, updateConfig := deploymentUpdateConfig(
-		`  wait_for_rollout = true`,
-		`
-  wait_for_rollout = false
-	wait_for {
-		condition {
-			type = "Available"
-			status = "True"
-		}
-	}
-`,
-		1,
-	)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: createConfig},
-			{Config: updateConfig},
-		},
-	})
-}
-
-func TestAccKubectl_WaitForFieldUpdate(t *testing.T) {
-	t.Parallel()
-
-	// replicas: 1 -> 2 forces metadata.generation to bump so
-	// status.observedGeneration can reach "2" — without a spec change the
-	// wait_for.field match for "2" would never succeed.
-	createConfig, updateConfig := deploymentUpdateConfig(
-		`  wait_for_rollout = true`,
-		`
-  wait_for_rollout = false
-	wait_for {
-		field {
-			key = "status.observedGeneration"
-			value = "2"
-		}
-	}
-`,
-		2,
-	)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{Config: createConfig},
-			{Config: updateConfig},
-		},
-	})
-}
-
-//func TestAccKubect_Debug(t *testing.T) {
-//	//language=hcl
-//	config := `
-//resource "kubectl_manifest" "test" {
-//	yaml_body = <<YAML
-//apiVersion: v1
-//kind: Secret
-//metadata:
-//  name: test-secret
-//stringData:
-//  var: "${formatdate("YYYYMMDDhhmmss", timestamp())}"
-//YAML
-//}
-//`
-//
-//	//start := time.Now()
-//	resource.Test(t, resource.TestCase{
-//		PreCheck:     func() { testAccPreCheck(t) },
-//		Providers:    testAccProviders,
-//		CheckDestroy: testAccCheckkubectlDestroy,
-//		Steps: []resource.TestStep{
-//			{
-//				Config: config,
-//				//todo: improve checking
-//			},
-//		},
-//	})
-//}
-
-func TestAccInconsistentPlanning(t *testing.T) {
-	t.Parallel()
-
-	// See https://github.com/alekc/terraform-provider-kubectl/pull/46
-	name := acctest.RandomWithPrefix("inconsistent-secret")
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "secret" {
-  yaml_body = <<EOF
-apiVersion: v1
-kind: Secret
-metadata:
-  name: %s
-stringData:
-  var: "${formatdate("YYYYMMDDhhmmss", timestamp())}"
-EOF
-}
-`, name)
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:             config,
-				ExpectNonEmptyPlan: true, // yaml_incluster is constantly different
-			},
-			{
-				// used to crash out on the second run
-				Config:             config,
-				ExpectNonEmptyPlan: true,
-			},
-		},
-	})
-}
-
-func TestAccKubectlUnknownNamespace(t *testing.T) {
-	t.Parallel()
-
-	// The namespace is hardcoded as one that will never exist, so no
-	// collision with parallel tests.
-	name := acctest.RandomWithPrefix("ingress-unknown-ns")
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	yaml_body = <<EOT
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: %s
-  namespace: this-doesnt-exist
-spec:
-  ingressClassName: "nginx"
-  rules:
-  - host: "*.example.com"
-    http:
-      paths:
-      - path: "/testpath"
-        pathType: "Prefix"
-        backend:
-          service:
-            name: test
-            port:
-              number: 80
-	EOT
-		}
-`, name)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config:      config,
-				ExpectError: regexp.MustCompile("\"this-doesnt-exist\" not found"),
-			},
-		},
-	})
-}
-
-func TestAccKubectlOverrideNamespace(t *testing.T) {
-	t.Parallel()
-
-	namespace := "dev-" + acctest.RandString(10)
-	yaml_body := `
-apiVersion: v1
-kind: Secret
-metadata:
-  name: mysecret
-  namespace: prod
-type: Opaque
-data:
-`
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "ns" {
-	yaml_body = <<EOT
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: %s
-    EOT
-}
-
-resource "kubectl_manifest" "test" {
-	depends_on = [kubectl_manifest.ns]
-    override_namespace = "%s"
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, namespace, namespace, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "namespace", namespace),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "override_namespace", namespace),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body_parsed", fmt.Sprintf(`apiVersion: v1
-data: (sensitive value)
-kind: Secret
-metadata:
-  name: mysecret
-  namespace: %s
-type: Opaque
-`, namespace)),
-				),
-			},
-		},
-	})
-}
-
-func TestAccKubectlSetNamespace(t *testing.T) {
-	t.Parallel()
-
-	namespace := "dev-" + acctest.RandString(10)
-	yaml_body := `
-apiVersion: v1
-kind: Secret
-metadata:
-  name: mysecret
-type: Opaque
-data:
-`
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "ns" {
-	yaml_body = <<EOT
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: %s
-    EOT
-}
-
-resource "kubectl_manifest" "test" {
-    depends_on = [kubectl_manifest.ns]
-    override_namespace = "%s"
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, namespace, namespace, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "id", "/api/v1/namespaces/"+namespace+"/secrets/mysecret"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "namespace", namespace),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "override_namespace", namespace),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body_parsed", fmt.Sprintf(`apiVersion: v1
-data: (sensitive value)
-kind: Secret
-metadata:
-  name: mysecret
-  namespace: %s
-type: Opaque
-`, namespace)),
-				),
-			},
-		},
-	})
-}
-
-func TestAccKubectlSetNamespace_nonnamespaced_resource(t *testing.T) {
-	t.Parallel()
-
-	namespace := "dev-" + acctest.RandString(10)
-	yaml_body := fmt.Sprintf(`
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: mysuperrole-%s
-rules:
-- apiGroups: [""]
-  resources: ["secrets"]
-  verbs: ["get", "watch", "list"]
-`, namespace)
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-    override_namespace = "%s"
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, namespace, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "namespace", namespace),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "override_namespace", namespace),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body_parsed", fmt.Sprintf(`apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: mysuperrole-%s
-  namespace: %s
-rules:
-- apiGroups:
-  - ""
-  resources:
-  - secrets
-  verbs:
-  - get
-  - watch
-  - list
-`, namespace, namespace)),
-				),
-			},
-		},
-	})
-}
-
-func TestAccKubectlSensitiveFields_secret(t *testing.T) {
-	t.Parallel()
-
-	name := acctest.RandomWithPrefix("sensitive-secret")
-	yaml_body := fmt.Sprintf(`
-apiVersion: v1
-kind: Secret
-metadata:
-  name: %s
-  namespace: default
-type: Opaque
-data:
-  USER_NAME: YWRtaW4=
-  PASSWORD: MWYyZDFlMmU2N2Rm
-`, name)
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "namespace", "default"),
-					resource.TestCheckNoResourceAttr("kubectl_manifest.test", "override_namespace"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body_parsed", fmt.Sprintf(`apiVersion: v1
-data: (sensitive value)
-kind: Secret
-metadata:
-  name: %s
-  namespace: default
-type: Opaque
-`, name)),
-				),
-			},
-		},
-	})
-}
-
-// ingressYAMLFixture renders an Ingress YAML fragment with a randomised name,
-// shared by the SensitiveFields_slice / _unknown_field and WithoutValidation
-// tests so each gets a unique cluster object.
-func ingressYAMLFixture(prefix string) (yaml, name string) {
-	name = acctest.RandomWithPrefix(prefix)
-	yaml = fmt.Sprintf(`
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: %s
-spec:
-  ingressClassName: "nginx"
-  rules:
-  - host: "*.example.com"
-    http:
-      paths:
-      - path: "/testpath"
-        pathType: "Prefix"
-        backend:
-          service:
-            name: test
-            port:
-              number: 80`, name)
-	return
-}
-
-func TestAccKubectlSensitiveFields_slice(t *testing.T) {
-	t.Parallel()
-
-	yaml_body, name := ingressYAMLFixture("sensitive-slice")
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-    sensitive_fields = [
-      "spec.rules",
-    ]
-
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body_parsed", fmt.Sprintf(`apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: %s
-spec:
-  ingressClassName: nginx
-  rules: (sensitive value)
-`, name)),
-				),
-			},
-		},
-	})
-}
-
-func TestAccKubectlSensitiveFields_unknown_field(t *testing.T) {
-	t.Parallel()
-
-	yaml_body, name := ingressYAMLFixture("sensitive-unknown")
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-    sensitive_fields = [
-      "spec.field.missing",
-    ]
-
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body_parsed", fmt.Sprintf(`apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: %s
-spec:
-  ingressClassName: nginx
-  rules:
-  - host: '*.example.com'
-    http:
-      paths:
-      - backend:
-          service:
-            name: test
-            port:
-              number: 80
-        path: /testpath
-        pathType: Prefix
-`, name)),
-				),
-			},
-		},
-	})
-}
-
-func TestAccKubectlWithoutValidation(t *testing.T) {
-	t.Parallel()
-
-	yaml_body, _ := ingressYAMLFixture("without-validation")
-
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-    validate_schema = false
-
-	yaml_body = <<EOT
-%s
-	EOT
-		}
-`, yaml_body)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: config,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "yaml_body", yaml_body+"\n"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "validate_schema", "false"),
-				),
-			},
-		},
-	})
-}
 
 func TestGetLiveManifestFilteredForUserProvidedOnly(t *testing.T) {
 	testCases := []struct {
@@ -1469,54 +463,10 @@ func TestGetLiveManifestFilteredForUserProvidedOnly(t *testing.T) {
 		})
 	}
 }
-
-func TestAccKubectlServerSideValidationFailure(t *testing.T) {
-	t.Parallel()
-
-	// The Ingress is invalid (backend service name violates DNS-1035), so
-	// the server-side validation rejects it before any cluster object is
-	// created — randomising the Ingress name is purely defensive against
-	// future changes to the test.
-	name := acctest.RandomWithPrefix("invalid-ingress")
-	config := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-  yaml_body = <<YAML
-apiVersion: networking.k8s.io/v1
-kind: Ingress
-metadata:
-  name: %s
-spec:
-  rules:
-    - host: "test-a.proxypile.tk"
-      http:
-        paths:
-          - path: /
-            pathType: Prefix
-            backend:
-              service:
-                name: nginx.test-a.svc.cluster.local
-                port:
-                  number: 8080
-YAML
-}
-`, name)
-	expectedError := regexp.MustCompile(".*Invalid value: \"nginx.test-a.svc.cluster.local\": a DNS-1035 label must consist of lower case alphanumeric characters.*")
-	resource.Test(t, resource.TestCase{
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				ExpectError: expectedError,
-				Config:      config,
-			},
-		},
-	})
-}
-
 func withAlteredField(manifest *yaml.Manifest, value interface{}, fields ...string) *yaml.Manifest {
 	_ = unstructured.SetNestedField(manifest.Raw.Object, value, fields...)
 	return manifest
 }
-
 func loadRealDeploymentManifest() *yaml.Manifest {
 	manifest, _ := yaml.ParseYAML(`
 apiVersion: apps/v1
@@ -1657,160 +607,4 @@ status:
 `)
 
 	return manifest
-}
-
-// TestAccKubectl_UpgradeApiVersion_InPlaceUpdate verifies that changing the
-// apiVersion with upgrade_api_version=true updates the resource in-place
-// (preserving its UID) rather than forcing a delete and recreate.
-func TestAccKubectl_UpgradeApiVersion_InPlaceUpdate(t *testing.T) {
-	t.Parallel()
-
-	name := "test-upgrade-api-version-" + acctest.RandString(10)
-
-	configV1 := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	upgrade_api_version = true
-	yaml_body = <<YAML
-apiVersion: autoscaling/v1
-kind: HorizontalPodAutoscaler
-metadata:
-  name: %s
-spec:
-  maxReplicas: 5
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: does-not-exist
-YAML
-}
-`, name)
-
-	configV2 := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	upgrade_api_version = true
-	yaml_body = <<YAML
-apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: %s
-spec:
-  maxReplicas: 5
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: does-not-exist
-YAML
-}
-`, name)
-
-	var uid string
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: configV1,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "api_version", "autoscaling/v1"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "upgrade_api_version", "true"),
-					// Capture the UID after initial creation
-					resource.TestCheckResourceAttrWith("kubectl_manifest.test", "uid", func(value string) error {
-						uid = value
-						return nil
-					}),
-				),
-			},
-			{
-				Config: configV2,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "api_version", "autoscaling/v2"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "upgrade_api_version", "true"),
-					// Verify UID is unchanged — proves in-place update, not recreate
-					resource.TestCheckResourceAttrWith("kubectl_manifest.test", "uid", func(value string) error {
-						if value != uid {
-							return fmt.Errorf("resource was recreated: UID changed from %s to %s", uid, value)
-						}
-						return nil
-					}),
-				),
-			},
-		},
-	})
-}
-
-// TestAccKubectl_UpgradeApiVersion_ForceNewByDefault verifies that changing the
-// apiVersion WITHOUT upgrade_api_version (default false) forces a delete and
-// recreate, resulting in a new UID.
-func TestAccKubectl_UpgradeApiVersion_ForceNewByDefault(t *testing.T) {
-	t.Parallel()
-
-	name := "test-upgrade-api-version-" + acctest.RandString(10)
-
-	configV1 := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	yaml_body = <<YAML
-apiVersion: autoscaling/v1
-kind: HorizontalPodAutoscaler
-metadata:
-  name: %s
-spec:
-  maxReplicas: 5
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: does-not-exist
-YAML
-}
-`, name)
-
-	configV2 := fmt.Sprintf(`
-resource "kubectl_manifest" "test" {
-	yaml_body = <<YAML
-apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: %s
-spec:
-  maxReplicas: 5
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: does-not-exist
-YAML
-}
-`, name)
-
-	var uid string
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckkubectlDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: configV1,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "api_version", "autoscaling/v1"),
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "upgrade_api_version", "false"),
-					resource.TestCheckResourceAttrWith("kubectl_manifest.test", "uid", func(value string) error {
-						uid = value
-						return nil
-					}),
-				),
-			},
-			{
-				Config: configV2,
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttr("kubectl_manifest.test", "api_version", "autoscaling/v2"),
-					// Verify UID changed — proves resource was recreated
-					resource.TestCheckResourceAttrWith("kubectl_manifest.test", "uid", func(value string) error {
-						if value == uid {
-							return fmt.Errorf("resource was NOT recreated: UID remained %s", uid)
-						}
-						return nil
-					}),
-				),
-			},
-		},
-	})
 }
